+++ src/Account.h
@@ -10,19 +10,19 @@ class Account {
  private:
   int account_id_;
   std::string name_;
-  float balance_;
+  double balance_;
   bool enabled_;
   bool student_;
 
  public:
-  Account(int,std::string,float,bool,bool);
+  Account(int,std::string,double,bool,bool);
   Account();
   int get_account_id();
   std::string get_name();
-  float get_balance();
+  double get_balance();
   bool is_enabled();
   bool is_student();
-  void set_balance(float);
+  void set_balance(double);
   void enable();
   void disable();
   void change_plan();
